[Part4d] Add user administration and authentication (exercises 4.15-4.22)

+++ part4/bloglist/controllers/blogs.js
@@ -1,32 +1,45 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const logger = require('../utils/logger')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1, id: 1 })
   response.json(blogs)
 })
 
+
 blogsRouter.post('/', async (request, response) => {
   const { title, author, url, likes } = request.body
+  const token = jwt.verify(request.token, process.env.SECRET)
+
+  const user = await User.findById(token.id)
   const blog = new Blog({
+    url,
     title,
     author,
-    url,
-    likes: likes || 0
+    likes: likes || 0,
+    user: user._id
   })
+
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
+
 blogsRouter.get('/:id', async (request, response) => {
-  const blog = await Blog.findById(request.params.id)
+  const blog = await Blog.findById(request.params.id).populate('user', { username: 1, name: 1, id: 1 })
   if (blog) {
     response.json(blog)
   } else {
     response.status(404).end()
   }
 })
+
 
 blogsRouter.put('/:id', async (request, response) => {
   const blog = {
@@ -37,7 +50,7 @@ blogsRouter.put('/:id', async (request, response) => {
   }
 
   const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true, runValidators: true })
-  
+
   if (updatedBlog) {
     response.json(updatedBlog)
   } else {
@@ -45,7 +58,15 @@ blogsRouter.put('/:id', async (request, response) => {
   }
 })
 
+
 blogsRouter.delete('/:id', async (request, response) => {
+  const token = jwt.verify(request.token, process.env.SECRET)
+
+  const blog = await Blog.findById(request.params.id)
+  if (token.id !== blog.user.toString()) {
+    return response.status(403).json({ error: 'unauthorized deletion' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
